Extract EditCommandParser from Parser class
